reorganize components into main component folder
