+++ src/components/AddPetForm/CategoryStep/CategoryStep.styled.jsx
